Add new SubnetCIDR method, and restore IPWithSubnetMask
In 7867fd91ac640dac63998fb9bae6d48da8451ddd, IPWithSubnetMask was
modified to return the CIDR of the subnet, and not the IP with the
subnet mask.  This was to fix an issue in the Docker CPI where Docker
stopped accepting subnet CIDRs with a specific IP attached.

Unfortunately, this broke the Warden CPI, which needs exactly that
functionality to provide the IP address to Garden.
(ref. https://github.com/cloudfoundry/garden/blob/74829d0a220114f627a733a0d4dfd4ac8d96c9f7/client.go#L107-L132)

This commit restores the previous behavior of IPWithSubnetMask and adds
a new SubnetCIDR method that returns the CIDR of the subnet.  This will
allow both the Docker and Warden CPIs to get the values that they need.

Co-authored-by: Chris Selzo <chris.selzo@broadcom.com>